fixed variable conversion error

+++ ev.py
@@ -1,6 +1,6 @@
 #from __future__ import division,print_function
 
-x=float(input("Enter a number? "))
+x=int(input("Enter a number? "))
 if x%2==0:	
 	print(x,"is even.")
 else:
